[mypyc] Fix is_native_ref_expr for class attrs (#18031)
This addresses https://github.com/mypyc/mypyc/issues/932 in which
in some cases before we were incorrectly not determining that the symbol
was native.

Because the problematic implementation of `is_native_ref_expr` is
present in multiple places,
the one in `builder.py` is rewritten to use the one in `mapper.py`.

Additional run tests are added to cover the case. Use of a dictionary
here prevents falling
under a secondary, more optimized codepath, that avoided the issue
entirely.