Do not prioritize ParamSpec signatures during overload resolution (#18033)
Closes #18027.

Var-args and star-args overloads are handled first to handle functions
like `zip` correctly in cases like `zip(*[[1],[2],[3]])`. That does not
seem to be necessary in case of `ParamSpec` overloads (or at least such
use case is much less obvious).

So this PR prevents `ParamSpec` overloads from floating up in the list
of overload targets.