Make deprecated Note a standard Error, disabled by default
While working on the relase of mypy 1.14 we noticed a large number of
notes for deprecated. Speaking with Jukka, he suggested we make this
disabled by default. And if it's disabled by default, having it as a
note is not as useful. We also don't have many stand alone notes. Most
notes are "attached" with an error.

This PR makes the deprecated error disabled by default, removes the flag
to report it as error and always reports it as error when enabled.