[mypyc] Fixing index variable in for-loop with builtins.enumerate. (#18202)
Fixes [mypyc/mypyc#1046](https://github.com/mypyc/mypyc/issues/1046)

This change fixes two problems:
1. The index variable was getting instantiated even while enumerating an
empty iterable.
2. After exiting the for-loop, the value of the index variable is off by
1 (see issue linked above).

This change fixes both problems by assigning the temporary register to
the index variable at the beginning of the for-loop body. Before this
change, this assignment was happening before the for-loop and at the end
of the for-loop body.