Make "deprecated" Note a standard Error, disabled by default (#18192)
While working on the relase of mypy 1.14 we noticed a large number of
notes for deprecated. Speaking with Jukka, he suggested we make this
disabled by default. And if it's disabled by default, having it as a
note is not as useful. We also don't have many stand alone notes. Most
notes are "attached" with an error.

This PR makes the deprecated error disabled by default and by default
reports it as error when enabled. `--report-deprecated-as-note` can
be used to report them as notes instead.