[misc] Properly format subheadings and code in gen_blog_post_html.py
Our Changelog is uses 3 number signs (#) for subheadings, not 4. This
should generate proper `<h3>` in the blog post.

Also, code blocks are generate with just `<pre>` tags. Let's also add
`<code>` tags so we can potentially apply syntax highlighting with js
like highlight.js in the future, if we want to.