[mypyc] Improve access to generated C on test failures and document this
Now the generated C files for the first mypyc run test failure in a
pytest session will be copied to the `.mypyc_test_output` directory,
and this will be indicated in the test output. This is a convenience
feature to help in the common scenario where all test failures have the
same root cause, so any single output is sufficient for debugging.

Document this and `--mypyc-showc`, which allows showing generated C
for every test failure. The latter is too verbose to be enabled by
default.