Better names of and more compatibility between ad hoc intersections of instances (#18506)
While working on #18433, we encountered [this
bug.](https://github.com/python/mypy/pull/18433#issuecomment-2583314142).
@ilevkivskyi identified [the underlying
problem](https://github.com/python/mypy/pull/18433#issuecomment-2585455830),
and we decided to try to reuse previously created ad hoc intersections
of instances instead of always creating new ones.

While working on this PR, I realised that reusing intersections requires
more complete names. Currently, module and type variable specifications
are not included, which could result in mistakes when using these names
as identifiers.

So, I switched to more complete names. Now, for example, `<subclass of
"A" and "A">` becomes `<subclass of "mod1.A[builtins.int]" and
"mod2.A">`. Hence, I had to adjust many existing test cases where
`reveal_type` is used.

`testReuseIntersectionForRepeatedIsinstanceCalls` confirms that the
mentioned bug is fixed.

`testIsInstanceAdHocIntersectionIncrementalNestedClass` and
`testIsInstanceAdHocIntersectionIncrementalUnions` are in a separate
commit. I think they are not really necessary, so we might prefer to
remove them. I added them originally because I had to adjust
`lookup_fully_qualified` a little. The change is very simple, but I
could not create a test case where it is not sufficient.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>