[stubgen] Include simple decorators in stub files (#18489)
Stubgen historically only included a selected number of decorators in
the generated stubs. I couldn't find the actual reason for it, however
it's likely fair to assume that decorator typing only started being
possible with PEP 612 thus most had been untyped previously.

As it's fairly simple to annotate decorators with `ParamSpec` now, it's
probably fair to include them in the stub file now.