Fix --warn-unreachable fall-through from ALWAYS_TRUE ifs
Currently blocks marked as unreachable in `reachability.py` through
semanal_pass1 are ignored by the `--warn-unreachable` flag through
a `block.is_unreachable` check. Namely, reachability marks the if
bodies and else-body as unreachable if it is not reachable on the
current platform or Python version.

This works for if-statements that have an else branch but if the
if-statement has a short-circuiting early return or raise on a
particular platform, then `reachability.py` places an empty else block
and marks it unreachable but has no effect on the fall-through code.

When `checker.py` correctly determines that the if-statement is always
True and has a return inside it, it correctly marks any code after
the if-statement as unreachable. This fix makes it so the code
after if-statements that are known to be always true is marked reachable
to surpress spurious `--warn-unreachable` warnings.