[mypyc] Add debug op (and builder helper) for printing str or Value (#18552)
It generates C code to print to stdout, but tries to preserve the error
state and not affect the code it's added to.

Added test for it, but also tested this by adding
`builder.debug_print(typ)` in `add_non_ext_class_attr_ann` and it prints
the class name.
It's also useful to use it like `builder.debug_print("MARKER")` and then
to search in the generated C code for MARKER. For more complex debugging
tasks, this is useful in finding your way around the generated C code
and quickly looking at the interesting part.

I saw that there's already a misc op `CPyDebug_Print`. I haven't seen it
used though. I think we can remove that one if this is proving useful.