[misc] fix blog post generator quirks (#18601)
While preparing the 1.15 release blog posts, I noticed a few quirks with
our maarkdown converter:

- We treat any `\n` followed by a capital letter as a new paragraph,
which can cause too many `<p>` tags to be inserted at times (a common
cause in this blog post was a line break followed by the word "PR").
- Using multiple consecutive backticks for an inline code section is
valid markdown (it's commonly used for strings where you need to include
single backticks, eg ``` ``a string with a single ` :)`` ```), but our
script was confused by this and generated lots of erroneous `<tt>`
sections where they didn't belong.
- Including a `#\d` in the middle of a word caused the script to assume
it was a PR that it should link. In this specific case, the changelog
contains several occurrences of `mypy_mypyc-wheels#<PR number>`, which
the script was stomping on.

This PR contains some minor tweaks for the blog post generation script
that attempt to address these quirks.