Allow lowered opts to be deleted in lowering trasnform (#18669)
This only works for simple initialization ops and ops where the return
value is ignored. This can be used for dummy init ops that are used to
give hints to data flow analysis about lifetimes of values, when
initialization is done via a pointer argument (e.g.
`init_my_struct(&reg)` in C).