Add an option to exclude everything in .gitignore (#18696)
Fixes https://github.com/python/mypy/issues/12505

This is (somewhat surprisingly) one of the most upvoted issues, and
looks like a simple thing to add. I essentially do what other tools do,
but optimize for how we work with sources discovery (to avoid
performance issues). I am making this opt-in for now, we can change this
later if needed.