Fix crash for callable with *args and suffix against Any (#18781)
Fixes https://github.com/python/mypy/issues/18780

Fix is trivial: handle a missing case. Note I re-use
`flatten_nested_tuples()` out of laziness. In theory, there should be at
most one level of nesting at this point, after which we should put an
assert (and IIRC we do something like this in other places). But I think
it is not worth the effort here, as this is a quite niche edge case
anyway.