Support positional and keyword-only arguments in stubdoc (#18762)
Currently the signature parsing logic fails when confronted with a `/`
or a `*`, rather than recognizing them as demarcating positional-only
and keyword-only arguments.

This patch supports parsing signatures with these features, but doesn't
pass this information along to the `ArgSig` or `FunctionSig` classes,
since the information would not be used anyway.