[mypyc] Detect more issues when generating annotated HTML (#18838)
Add new heuristics to detect various potential performance issues when
using `-a foo.html`.
The issues include various generic (non-specialized) operations, uses of
`functools` and
`itertools`, and slow `isinstance` checks that use runtime-checkable
protocols.

Implement a mypy AST visitor that is used to detect some issues that
would be harder to detect
when analyzing the generated IR.

Support annotation priorities so that if multiple annotations are
generated for a line, only the
highest-priority ones are shown. This is a bit crude but useful, since
often multiple heuristics
are triggered by some inefficient code, and duplicate annotations would
be verbose and
sometimes confusing.