Handle union types when binding self (#18867)
Currently we only bind self if the type is callable, but we actually
should do this for all callable items in a union.

This use case is probably quite niche (since adding an annotation makes
a variable an instance variable, and we rarely infer unions). I found it
when looking at `checkmember`-related issues it was easy to handle it. I
also use this opportunity to refactor and add comments to
`analyze_var()`.