[mypyc] Add hidden flag to skip the generation of C files (#18955)
This can be useful when debugging mypyc issues. For example, you can
manually add some debug prints to the generated C and rerun mypyc
with `--skip-c-gen`. Now mypyc will build the C code again, with your
manual changes included (this assumes everything else is the same as
during the previous run).

I'm not planning to advertise this as an end-user feature.