[mypyc] Fix spilling values with overlapping error values (#18961)
Update spilling transform for async and generator functions to not
require a defined attribute bitfield. We now treat temporary spilled
values with overlapping error values as always defined. I'm not sure if
this would be safe to do for reference counted values, so we only do
this in cases where it's clearly fine.