Check superclass compatibility of untyped methods if `--check-untyped-defs` is set (#18970)
This PR enables superclass compatibility checks for untyped methods when
`--check-untyped-defs` is set. IMO this behavior is correct as
`--check-untyped-defs` is essentially "treat everything as if there were
`: Any` annotations on all arguments", hence checking arg count and
names is sound.

This PR, however, allows `@override` on classes that have Any fallback
as those are often coming from unfollowed imports. This PR started as an
attempt to reject `@override` on untyped defs not found in superclass,
but I think it's better to just run all compatibility checks if the flag
is enabled.