[mypyc] Refactor IR building for generator functions (#19008)
The code was pretty hard to follow, since there were many conditional
code paths in a very long function that handles both normal functions,
nested functions and generators.

Move much of the generator-specific code to a helper function. This
required moving some functionality to helper functions to avoid code
duplication. Also pass a function as an argument to avoid a dependency
cycle.

This is quite a big refactoring, and it's easier to follow by looking at
the individual commits in this PR.