Add some functionality to misc/perf_compare.py (#18471)
While working on determining the performance impact of #17875, I
discovered misc/perf_compare.py. I extended its functionality slightly,
introducing an -r flag to check a foreign repo instead of mypy itself. I
also added a --dont-setup flag, to save time on recompiling mypy. I
added several helpful printouts: the standard deviations of the samples,
a line at the bottom listing the total time taken by the whole
benchmarking program. I improved the cli documentation with the
docstring, better descriptions, and an epilog reminding you to probably
specify master.

As a caveat, I should flag: when running the benchmark on the foreign
repo, it typically spits out a lot of errors, presumably because the
libraries aren't installed. That makes sense to me, though, and seems
fine.

Another caveat I'll flag on the off-chance it's important: on my
machine, this script seems constitutionally incapable of deleting the
tmpdirs it makes. It fails with a permissions error. But it was like
that when I got it.