Add --strict-bytes to --strict (#19049)
This is a check that ensures static correctness, so it is useful to have
in --strict. Unlike making this the default behavior eventually in 2.0,
which we are also planning to do, it can be added to --strict
immediately due to --strict having looser backwards-compatibility
requirements (or so I interpret --strict's
[documentation](https://mypy.readthedocs.io/en/stable/command_line.html#cmdoption-mypy-strict)).
This PR also includes tests, for --strict and also for no flags.