Update dmypy/client.py:  Enable ANSI color codes for windows cmd (#19088)
I still use windows cmd, and the color codes emitted by dmypy do not
work on there. instead printing a bunch of codes like ←[37m
←[39;49;00ms. However, for whatever reason you can fix this simply by
calling os.system("") once. (The main mypy program works fine,
presumably because it makes an os system call somewhere before it
prints.)

I did not write a test of this, as that seems difficult and unnecessary.
Instead, I manually tested it, and it worked great.