Move dataclass kw_only fields to the end of the signature (#19018)
Fixes #19017. Fixes #17731.

This is a rather naive change: python does that at runtime. `kw_only`
args can be in any order, and non-kwonly args should remain sorted as-is
(stable sort). I don't understand why this was only done in presence of
a parent dataclass - AFAIC kwonly fields work that way since `kw_only`
was introduced in py3.10.

The test I changed was invalid and asserted a false positive to the best
of my knowledge.