Erase stray typevars in functools.partial generic (#18954)
Fixes #18953. Fixes #15215. Refs #17461.

When the function passed to `partial` is generic and has generic params
in the return type, we must erase them, otherwise they become orphan and
cannot be used later. This only applies to `partial[...]` generic param
and not to the underlying "exact" callable stored internally as the
latter remains generic.

The ultimate fix would be to implement #17620 so that we stop caring
about `partial[...]` generic param, but this should improve usability
(but causes false negatives).