[mypyc] Support overriding the group name used in output files
By default, when compiling more than one file, a shared library
is generated with a file derived from the sha of compiled modules
(i.e. the group name). This is also used in the names of generated
.c and .h files. Add experimental `group_name` argument to `mypycify`
that allows overriding this. This can be useful when integrating
mypyc to a build system, as this makes the names of output files
more predictable.