[mypyc] Derive .c file name from full module name if using multi_file (#19278)
Don't shorten the module name prefixes, so that the names of output
files are more predictable. This can help if integrating with build
systems, and it arguably also makes it easier to inspect the output
manually.

Now the file name could be `__native_pkg___mod.c` instead of
`__native_mod.c` for the module `pkg.mod`.