Use union of current context and left side for right side narrowing of binary ops (#19249)
Fixes #12001. Fixes #6898. Fixes #15368. Improves #17790 and #11508.

When encountering `a {and,or} b`, we used to use `a` as primary context
for `b` inference. This results in weird errors when `a` and `b` are
completely unrelated, and in many cases return type/assignment type
context can do much better. Inferring to union should be harmless in
most cases, so use union of `a` and current context instead.