[mypyc] Test cases and compilation error for mypyc-try-finally-await
Any await that causes a context switch inside a finally block
swallows any exception (re-)raised in the preceding try or except
blocks. As the exception 'never happened', this also changes
control flow.

This commit adds several tests (which fail) for this bug, and
triggers a compiler error if this pattern is detected in the code.
`# type: ignore[mypyc-try-finally-await, unused-ignore]` can be
used on the try line to bypass the error.

This also newly causes the testAsyncReturn test to fail, as it
should.

See mypyc/mypyc#1114