Metaclass conflict check improvements (#17682)
This PR fixes some points of #14033:
 * Give metaclass errors to their own error code (I chose `metaclass`).
* Document shortcomings of and workarounds for mypy's metaclass
handling.

I didn't attempt to fix that mypy follows the logic for determining the
metaclass as documented whereas it should follow what the interpreter is
actually doing
(https://github.com/python/mypy/issues/14033#issuecomment-1314025562). I
think such a change is better kept as a separate PR, which is why I
don't want to close the issue with this PR.

Fixes: https://github.com/python/mypy/issues/14033

---------

Co-authored-by: hauntsaninja <hauntsaninja@gmail.com>