[mypyc] Add support for C string literals in the IR (#19383)
Previously only Python str and bytes literals were supported, but
sometimes we want zero-terminated C string literals instead. They
don't need to be allocated from the heap and are usually stored in a
read-only data section, so they are more efficient in some use cases.

These will be useful for a feature I'm working on.